Imported upstream version '1.13.18' of 'upstream'

+++ src/rviz/default_plugin/path_display.h
@@ -33,7 +33,7 @@
 
 #include <nav_msgs/Path.h>
 
-#include <rviz/message_filter_display.h>
+#include "rviz/message_filter_display.h"
 #include <rviz/ogre_helpers/arrow.h>
 #include <rviz/ogre_helpers/axes.h>
 
@@ -85,8 +85,8 @@ private Q_SLOTS:
 
 private:
   void destroyObjects();
-  void allocateArrowVector(std::vector<rviz::Arrow*>& arrow_vect, size_t num);
-  void allocateAxesVector(std::vector<rviz::Axes*>& axes_vect, size_t num);
+  void allocateArrowVector(std::vector<rviz::Arrow*>& arrow_vect, int num);
+  void allocateAxesVector(std::vector<rviz::Axes*>& axes_vect, int num);
   void destroyPoseAxesChain();
   void destroyPoseArrowChain();
